requirements.txt
streamlit==1.32.0
snowflake-connector-python==3.6.1
kafka-python==2.0.2
requests==2.31.0
beautifulsoup4==4.12.3
yfinance==0.2.36
pandas==2.2.1
matplotlib==3.8.3
textblob==0.17.1
nltk==3.8.1
statsmodels==0.14.1
pmdarima==2.0.4
